requirements.txt
# Web framework
fastapi[standard]

# Database
sqlalchemy
psycopg2-binary

# Environment
python-dotenv

# LangChain suite (all verified by imports)
langchain-core
langchain-community
langchain-chroma
langchain-cohere
langchain-ollama
langchain-classic
langchain-huggingface

# Hugging face
huggingface-hub

# Vector DB (backend for langchain-chroma)
chromadb

# Document processing
unstructured[pdf]

# Embeddings
sentence-transformers

# Data validation
pydantic

# Encryption
bcrypt==4.3.0
